overlaps warning processed

+++ src/main.py
@@ -22,24 +22,33 @@ def preview(api: sly.Api, task_id, context, state, app_logger):
 
     api.task.set_fields(task_id, fields)
 
-    process_video(sly_progress, state, is_preview=True)
-    sly_progress.refresh_params('Uploading video', 1)
+    rc, _ = process_video(sly_progress, state, is_preview=True)
 
-    video_path = os.path.join(app.data_dir, './preview.mp4')
-    remote_video_path = os.path.join(f"/SLYvSynth/{task_id}", "preview.mp4")
-    if api.file.exists(team_id, remote_video_path):
-        api.file.remove(team_id, remote_video_path)
+    if rc == 0:
+        sly_progress.refresh_params('Uploading video', 1)
 
-    # file_info = api.file.upload(team_id, video_path, remote_video_path, progress_cb=sly_progress.pbar.iters_done_report)
-    file_info = api.file.upload(team_id, video_path, remote_video_path)
+        video_path = os.path.join(app.data_dir, './preview.mp4')
+        remote_video_path = os.path.join(f"/SLYvSynth/{task_id}", "preview.mp4")
+        if api.file.exists(team_id, remote_video_path):
+            api.file.remove(team_id, remote_video_path)
 
-    sly_progress.next_step()
+        # file_info = api.file.upload(team_id, video_path, remote_video_path, progress_cb=sly_progress.pbar.iters_done_report)
+        file_info = api.file.upload(team_id, video_path, remote_video_path)
 
-    fields = [
-        {"field": "data.videoUrl", "payload": file_info.full_storage_url},
-        {"field": "state.previewLoading", "payload": False},
-    ]
-    api.task.set_fields(task_id, fields)
+        sly_progress.next_step()
+
+        fields = [
+            {"field": "data.videoUrl", "payload": file_info.full_storage_url},
+            {"field": "state.previewLoading", "payload": False},
+        ]
+        api.task.set_fields(task_id, fields)
+    else:  # @TODO: process errors
+        fields = [
+            {"field": "data.videoUrl", "payload": None},
+            {"field": "state.previewLoading", "payload": False},
+        ]
+        api.task.set_fields(task_id, fields)
+
 
 
 def reset_output_headers_by_state(state):
@@ -57,21 +66,30 @@ def synthesize(api: sly.Api, task_id, context, state, app_logger):
 
     reset_output_headers_by_state(state)
 
-    res_project_id = process_video(sly_progress, state, is_preview=False)
+    rc, res_project_id = process_video(sly_progress, state, is_preview=False)
 
-    res_project = api.project.get_info_by_id(res_project_id)  # load full project info
+    if rc == 0:
+        res_project = api.project.get_info_by_id(res_project_id)  # load full project info
 
-    fields = [
-        {"field": "state.step5Loading", "payload": False},
-        {"field": "state.done5", "payload": True},
+        fields = [
+            {"field": "state.step5Loading", "payload": False},
+            {"field": "state.done5", "payload": True},
 
-        {"field": "data.dstProjectId", "payload": res_project.id},
-        {"field": "data.dstProjectName", "payload": res_project.name},
-        {"field": "data.dstProjectPreviewUrl",
-         "payload": api.image.preview_url(res_project.reference_image_url, 100, 100)},
-    ]
-    api.task.set_fields(task_id, fields)
-    api.task.set_output_project(task_id, res_project.id, res_project.name)
+            {"field": "data.dstProjectId", "payload": res_project.id},
+            {"field": "data.dstProjectName", "payload": res_project.name},
+            {"field": "data.dstProjectPreviewUrl",
+             "payload": api.image.preview_url(res_project.reference_image_url, 100, 100)},
+        ]
+        api.task.set_fields(task_id, fields)
+        api.task.set_output_project(task_id, res_project.id, res_project.name)
+
+    else:
+        fields = [
+            {"field": "state.step5Loading", "payload": False},
+        ]
+        api.task.set_fields(task_id, fields)
+
+
     # app.stop()
